优化代码

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -16,6 +16,7 @@ use App\Handlers\Level;
 
 class CategoryController extends BaseController
 {
+	//分类列表
 	public function index(Level $level)
 	{
 		$list = Category::orderBy('sort', 'desc')->get();
@@ -24,6 +25,7 @@ class CategoryController extends BaseController
 		return view('admin.category.index', ['list'=> $list]);
 	}
 
+	//新增界面
 	public function create(Level $level)
 	{
 		$list = Category::orderBy('id', 'asc')->get();
@@ -32,6 +34,7 @@ class CategoryController extends BaseController
 		return view('admin.category.create', ['list'=> $list]);
 	}
 
+	//新增操作
 	public function store(CategoryRequest $request)
 	{
 		$category = Category::create([
@@ -43,11 +46,10 @@ class CategoryController extends BaseController
 			'sort' => $request->sort
 		]);
 
-		session()->flash('success', '创建成功');
-
-		return redirect()->route('admin.category.index');
+		return redirect()->route('admin.category.index')->with('success', '创建成功');
 	}
 
+	//编辑界面
 	public function edit(Category $category, Level $level)
 	{
 		$list = Category::orderBy('id', 'asc')->get();
@@ -56,6 +58,7 @@ class CategoryController extends BaseController
 		return view('admin.category.create', ['list'=> $list, 'category'=> $category]);
 	}
 
+	//编辑操作
 	public function update(CategoryRequest $request, Category $category, Level $level)
 	{
 		$data = $request->all();
@@ -66,19 +69,23 @@ class CategoryController extends BaseController
 		if(in_array($data['pid'], $childs_id_arr)){
 			return redirect()->back()->with('danger', '父级分类不能选取子级作为父级');
 		}
-
 		$category->update($data);
+
 		return redirect()->back()->with('success', '编辑成功');
 	}
 
+	//删除处理
 	public function destroy(Category $category)
 	{
 		if($category->hasChild()){
 			return redirect()->back()->with('danger', '请先删除子级分类');
 		}
-
+		if($category->posts->count()){
+			return redirect()->back()->with('danger', '请先删除分类下的文章');
+		}
+		
 		$category->delete();
-
+		
 		return redirect()->route('admin.category.index')->with('success', '删除成功');
 	}
 }
